requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pypdf>=3.0.0
python-pptx>=0.6.21
requests>=2.31.0
pytest>=7.0.0
pytest-cov>=4.1.0 